requirements.txt
Django==1.10
psycopg2==2.6.2

# Django rest framework
djangorestframework==3.1.3
djangorestframework-hstore==1.1

# Crontab
django-crontab==0.7.0

# Test
mock==1.0.1
nose==1.3.1
factory-boy==2.4.1